Lowercase user inputs in object assignment.
Previously the user inputs for city, month, and day
only used the .lower() function on the while loop when comparing
the user input to the set of valid values.  Moved the .lower() method
to the object assignment so calling the user inputs for city,
month, and day will result in uniform lowercase strings.